policyboom/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "policyboom"
version = "0.1.0"
description = "Enterprise legal risk intelligence CLI - analyze Terms of Service and Privacy Policies with multi-domain scanning"
authors = [{name = "PolicyBoom Team"}]
readme = "README.md"
requires-python = ">=3.11"
keywords = ["legal", "privacy", "terms-of-service", "compliance", "analysis", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "lxml-html-clean>=0.1.1",
    "readability-lxml>=0.8.1",
    "tldextract>=5.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
policyboom = "policyboom.cli:main"

[project.urls]
Homepage = "https://github.com/policyboom/policyboom"
Documentation = "https://policyboom.dev"
Repository = "https://github.com/policyboom/policyboom"
Issues = "https://github.com/policyboom/policyboom/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["policyboom*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
